requirements.txt
customtkinter==5.2.2
pystray==0.19.5
pillow==11.2.1
cairosvg==2.7.1
pywin32==310 